+++ Source/errors/early-termination-signal.js
