+++ Source/selection-manager.js
