+++ Source/status-bar/status-bar-service.js
