+++ Source/status-bar/status-bar-view.js
