+++ Source/utils/editor-finders.js
