+++ Source/utils/escape-reg-exp.js
