+++ Source/utils/is-word-selected.js
